refactor: extract swap logic into hook
